Refactor Navigation for typescript
