--- src/components/Navigation.css
@@ -1,4 +1,0 @@
-.navbar-brand img {
-  width: 30px;
-  height: 30px;
-}
